Alternative way of vendor decorator
