+++ db/seeds/002_spree.rb
@@ -3,3 +3,21 @@ Zabka_Winogrady = Spree::Vendor.create(
     about_us: "Winogrady 61",
     postcode: "61-659"
 )
+
+Zabka_Gorczyn = Spree::Vendor.create(
+    name: "Zabka Gorczyn",
+    about_us: "Głogowska 173a",
+    postcode: "60-126"
+)
+
+Zabka_Centrum = Spree::Vendor.create(
+    name: "Zabka Centrum",
+    about_us: "Wrocławska 23",
+    postcode: "61-838"
+)
+
+Zabka_Junikowo = Spree::Vendor.create(
+    name: "Zabka Junikowo",
+    about_us: "Malwowa 103a",
+    postcode: "60-175"
+)
